refactor(users): exportar Role en la superficie pública de users
Los endpoints de accounts y de system importaban Role de
users.services.permissions, entrando en las tripas de otro módulo. Se
hacía así porque pedírselo al paquete users durante su propia
inicialización rompía el ciclo de imports; con el ciclo resuelto, users
puede exportarlo como el resto de sus decoradores de permisos.

Actualiza las entradas de KNOWN_VIOLATIONS que el cambio resuelve o
desplaza.

Refs #582
